+++ Assets/_Quarantine/Code/Items/Configuration/ItemConfiguration.cs
@@ -5,9 +5,9 @@ namespace _Quarantine.Code.Items.Configuration
     [CreateAssetMenu(fileName = "Base Item Configuration", menuName = "Inventory System/Base Item Configuration")]
     public class ItemConfiguration : ScriptableObject
     {
-        [SerializeField] private float _durability;
+        [SerializeField] private int _durability;
 
         public string ID => name;
-        public float Durability => _durability;
+        public int Durability => _durability;
     }
 }
